+++ ext_localconf.php
@@ -31,7 +31,7 @@ $iconsFolder         = \T3v\T3vCore\Utility\ExtensionUtility::iconsFolder($exten
   ]
 );
 
-// === Page TSconfig ===
+// === Frontend ===
 
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPageTSConfig("<INCLUDE_TYPOSCRIPT: source=\"{$configurationFolder}/TSconfig.ts\">");